+++ packages/dataviews/src/field-types/date.tsx
@@ -10,11 +10,9 @@ import type {
 	DataViewRenderFieldProps,
 	Field,
 	FormatDate,
-	NormalizedField,
-	Operator,
-	Rules,
 	SortDirection,
 } from '../types';
+import type { FieldType } from '../types/private';
 import RenderFromElements from './utils/render-from-elements';
 import {
 	OPERATOR_ON,
@@ -28,13 +26,8 @@ import {
 	OPERATOR_BETWEEN,
 	DAYS_OF_WEEK,
 } from '../constants';
-import { getControl } from '../dataform-controls';
-import hasElements from './utils/has-elements';
-import getValueFromId from './utils/get-value-from-id';
-import setValueFromId from './utils/set-value-from-id';
-import getFilterBy from './utils/get-filter-by';
 
-function getFormat( field: Field< any > ): Required< FormatDate > {
+function getFormat< Item >( field: Field< Item > ): Required< FormatDate > {
 	return {
 		date:
 			field.format?.date !== undefined &&
@@ -74,26 +67,25 @@ function render( { item, field }: DataViewRenderFieldProps< any > ) {
 	return dateI18n( format.date, getDate( value ) );
 }
 
-export default function normalizeField< Item >(
-	field: Field< Item >
-): NormalizedField< Item > {
-	const getValue = field.getValue || getValueFromId( field.id );
-	const setValue = field.setValue || setValueFromId( field.id );
-	const isValid: Rules< Item > = {
+const sort = ( a: any, b: any, direction: SortDirection ) => {
+	const timeA = new Date( a ).getTime();
+	const timeB = new Date( b ).getTime();
+
+	return direction === 'asc' ? timeA - timeB : timeB - timeA;
+};
+
+export default {
+	type: 'date',
+	render,
+	Edit: 'date',
+	sort,
+	isValid: {
 		elements: true,
 		custom: () => null,
-	};
-
-	const sort = ( a: Item, b: Item, direction: SortDirection ) => {
-		const valueA = getValue( { item: a } );
-		const valueB = getValue( { item: b } );
-		const timeA = new Date( valueA ).getTime();
-		const timeB = new Date( valueB ).getTime();
-
-		return direction === 'asc' ? timeA - timeB : timeB - timeA;
-	};
-
-	const defaultOperators: Operator[] = [
+	},
+	enableSorting: true,
+	enableGlobalSearch: false,
+	defaultOperators: [
 		OPERATOR_ON,
 		OPERATOR_NOT_ON,
 		OPERATOR_BEFORE,
@@ -103,9 +95,8 @@ export default function normalizeField< Item >(
 		OPERATOR_IN_THE_PAST,
 		OPERATOR_OVER,
 		OPERATOR_BETWEEN,
-	];
-
-	const validOperators: Operator[] = [
+	],
+	validOperators: [
 		OPERATOR_ON,
 		OPERATOR_NOT_ON,
 		OPERATOR_BEFORE,
@@ -115,33 +106,6 @@ export default function normalizeField< Item >(
 		OPERATOR_IN_THE_PAST,
 		OPERATOR_OVER,
 		OPERATOR_BETWEEN,
-	];
-
-	return {
-		id: field.id,
-		type: 'date',
-		label: field.label || field.id,
-		header: field.header || field.label || field.id,
-		description: field.description,
-		placeholder: field.placeholder,
-		getValue,
-		setValue,
-		elements: field.elements,
-		getElements: field.getElements,
-		hasElements: hasElements( field ),
-		render: field.render ?? render,
-		Edit: getControl( field, 'date' ),
-		sort: field.sort ?? sort,
-		isValid: {
-			...isValid,
-			...field.isValid,
-		},
-		isVisible: field.isVisible,
-		enableSorting: field.enableSorting ?? true,
-		enableGlobalSearch: field.enableGlobalSearch ?? false,
-		enableHiding: field.enableHiding ?? true,
-		readOnly: field.readOnly ?? false,
-		filterBy: getFilterBy( field, defaultOperators, validOperators ),
-		format: getFormat( field ),
-	};
-}
+	],
+	getFormat,
+} satisfies FieldType< any >;